Update the Student Details

+++ db.py
@@ -155,31 +155,36 @@ def publishdetails(exam_details, exam_name, exam_code, class_name):
 
 def students_record(students):
     connection = db_connection()
-    with connection.cursor() as cursor:
-        cursor.execute("TRUNCATE TABLE students_information")
-        connection.commit()
-        query = """
-            INSERT INTO students_information (student_name, student_id, father_name, mother_name,
-            student_age, fatherphone, motherphone, address, place, image_url )
-            VALUES (%s, %s, %s, %s, %s, %s, %s, %s, %s, %s)"""
-        values = [
-            (
-                student["student_name"],
-                student["student_id"],
-                student["father_name"],
-                student["mother_name"],
-                student["student_age"],
-                student["fatherphone"],
-                student["motherphone"],
-                student["address"],
-                student["place"],
-                student["image_url"],
-            )
-            for student in students
-        ]
-        cursor.executemany(query, values)
-        connection.commit()
-
+    if connection:
+        try:
+            with connection.cursor() as cursor:
+                cursor.execute("TRUNCATE TABLE students_information")
+                connection.commit()
+                query = """
+                INSERT INTO students_information (student_name, student_id, father_name, mother_name,
+                student_age, fatherphone, motherphone, address, place, image_url )
+                VALUES (%s, %s, %s, %s, %s, %s, %s, %s, %s, %s)"""
+                values = [
+                    (
+                        student["student_name"],
+                        student["student_id"],
+                        student["father_name"],
+                        student["mother_name"],
+                        student["student_age"],
+                        student["fatherphone"],
+                        student["motherphone"],
+                        student["address"],
+                        student["place"],
+                        student["image_url"],
+                    )
+                    for student in students
+                ]
+                cursor.executemany(query, values)
+                connection.commit()
+        except Exception as e:
+            print(f"Error:{e}")
+        finally:
+            connection.close()
 
 # clear exams data
 # load exams data from json
@@ -187,34 +192,81 @@ def students_record(students):
 
 def exams_record(exam_lists):
     connection = db_connection()
-    with connection.cursor() as cursor:
-        # Delete the previous values in the database
-        cursor.execute(" TRUNCATE TABLE exam_table;")
-        connection.commit()
-        # Insert the New values in the Database
-        exam_query = """ INSERT INTO exam_table (exam_name, exam_code, class_name) VALUES (%s, %s, %s)"""
-        exam_values = [
-            (item["exam_name"], item["exam_code"], item["class_name"])
-            for item in exam_lists.get("exam_table", [])
-        ]
-        cursor.executemany(exam_query, exam_values)
-        connection.commit()
+    if connection:
+        try:
+            with connection.cursor() as cursor: 
+                #Delete the previous values in the database
+                cursor.execute(" TRUNCATE TABLE exam_table;")
+                connection.commit()
+                # Insert the New values in the Database
+                exam_query = """ INSERT INTO exam_table (exam_name, exam_code, class_name) VALUES (%s, %s, %s)"""
+                exam_values = [
+                    (item["exam_name"], item["exam_code"], item["class_name"])
+                    for item in exam_lists.get("exam_table", [])
+                ]
+                cursor.executemany(exam_query, exam_values)
+                connection.commit()
 
-        # Delete the Previous values in the Database
-        cursor.execute(" TRUNCATE TABLE exam_subjects_table; ")
-        connection.commit()
+                # Delete the Previous values in the Database
+                cursor.execute(" TRUNCATE TABLE exam_subjects_table; ")
+                connection.commit()
 
-        # Insert the New values in the Database
-        exam_subjects_query = """ INSERT INTO exam_subjects_table (exam_code, subject_name, exam_date, exam_time, marks) VALUES (%s, %s, %s, %s, %s) """
-        exam_subjects_value = [
-            (
-                item["exam_code"],
-                item["subject_name"],
-                item["exam_date"],
-                item["exam_time"],
-                item["marks"],
-            )
-            for item in exam_lists.get("exam_subjects_table", [])
-        ]
-        cursor.executemany(exam_subjects_query, exam_subjects_value)
-        connection.commit()
+                # Insert the New values in the Database
+                exam_subjects_query = """ INSERT INTO exam_subjects_table (exam_code, subject_name, exam_date, exam_time, marks) VALUES (%s, %s, %s, %s, %s) """
+                exam_subjects_value = [
+                    (
+                        item["exam_code"],
+                        item["subject_name"],
+                        item["exam_date"],
+                        item["exam_time"],
+                        item["marks"],
+                    )
+                    for item in exam_lists.get("exam_subjects_table", [])
+                ]
+                cursor.executemany(exam_subjects_query, exam_subjects_value)
+                connection.commit()
+        except Exception as e:
+            print(f"Error:{e}")
+        finally:
+            connection.close()
+
+
+def update(student):
+    connection = db_connection()
+    if connection:
+        try:
+            with connection.cursor() as cursor:
+
+                query = """ UPDATE students_information
+                SET 
+                    student_name=%s,
+                    father_name=%s,
+                    mother_name=%s,
+                    student_age=%s,
+                    fatherphone=%s,
+                    motherphone=%s,
+                    place=%s,
+                    address=%s
+                    
+                    WHERE student_id=%s  """
+
+                values=(
+                    student['student_name'],
+                    student['father_name'],
+                    student['mother_name'],
+                    student['student_age'],
+                    student['father_phone'],
+                    student['mother_phone'],
+                    student['place'],
+                    student['address'],
+                    student['student_id']
+                    )
+                
+                cursor.execute(query,values)
+                connection.commit()
+                print("Successfully Values are Updated")
+
+        except Exception as e:
+            print(f"Error : {e}")
+        finally:
+            connection.close()
